CXXCBC-830: Unstick Jenkins matrix, shrink Windows wall-clock, survive fork(child) (#952)
The cxx-scripted-build-pipeline regressed on every Linux stage after
CXXCBC-777 / CXXCBC-778 added gRPC and protobuf as FetchContent
dependencies, the Windows stage's wall-clock grew to ~2 h, and once
the build matrix turned green the integration suite then surfaced a
null-pointer dereference in the fork example plus a stale env-vs-runtime
gate in one observability test. Seven consecutive runs (5223, 5224,
5226, 5227, 5229, 5232, 5238) peeled off cascading failures at
successive layers; CXXCBC-830 enumerates the observed symptoms.

Four groups of code-side fixes:

* Third-party wiring. cmake/ThirdPartyDependencies.cmake sets
  CURL_ENABLE_EXPORT_TARGET=OFF and CURL_DISABLE_SRP=ON before the
  OpenTelemetry cpmaddpackage so curl can configure and compile
  against our static BoringSSL on Linux hosts without libcurl-dev.
  cmake/Cache.cmake stops propagating find_program's
  CACHE_BINARY-NOTFOUND sentinel into CMAKE_*_COMPILER_LAUNCHER.
  cmake/CompilerWarnings.cmake confines four clang-only -Wno-*
  flags to the CLANG_WARNINGS list so GCC's -Werror is happy.

* Toolchain strictness. core/protocol/cmd_mutate_in.hxx,
  test/test_integration_tracer.cxx, and tools/sysinfo.cxx each get
  a single-line change to satisfy GCC 8/9/10's stricter
  -Wconversion, -Wpedantic, and -Wsign-compare respectively.
  test/test_unit_mcbp_session.cxx moves a constexpr int into the
  lambda body so MSVC accepts it without making clang complain
  about an unused capture.

* Windows throughput. CMakeLists.txt adds add_compile_options(/MP)
  for MSVC at top scope, scoped to C/CXX via $<COMPILE_LANGUAGE> so
  it does not reach BoringSSL's NASM-assembled x86 .asm sources.
  All FetchContent dependencies inherit the flag. bin/build-tests.rb
  and .github/workflows/build.yml let --parallel default to
  Etc.nprocessors instead of capping at 1 or 2.
  cmake/GrpcProtobuf.cmake marks six unused gRPC subtargets
  EXCLUDE_FROM_ALL so grpc_unsecure et al. stop landing in ALL_BUILD.

* Runtime correctness. core/cluster.cxx stops std::move()'ing the
  tracer_, meter_, app_telemetry_meter_, app_telemetry_reporter_, and
  orphan_reporter_ members out during close(): the public-API
  cluster::notify_fork(fork_event::child) recreates the outer impl,
  but pre-fork bucket/collection refs keep the inner core impl alive
  through their shared_ptr, so observability fields zeroed by close()
  produced null derefs on the next call. Stop in place instead, with
  a comment block above close() explaining the rule for future
  maintainers (members whose public accessor does not check stopped_
  must not be moved out on close). test/test_integration_wrapper_tracer.cxx
  changes its version gate from integration.ctx.version (sourced from
  TEST_SERVER_VERSION env var, defaults to 6.6.0 when unset and the
  Jenkins pipeline does not export it) to integration.cluster_version()
  (runtime-derived from /pools/default), aligning with every other
  test in the suite.

The Jenkins pipeline definition still passes /m:4 explicitly; that
external one-line change is documented in jenkins-disk-defense.md.
Local reconfigure + SDK rebuild + affected-test rebuild under GCC 16
are clean; matrix validation comes from the next
cxx-scripted-build-pipeline run.

+++ tools/sysinfo.cxx
@@ -310,7 +310,7 @@ public:
     std::signal(SIGINT, sigint_handler);
     std::signal(SIGTERM, sigint_handler);
 
-    if (pid_ > std::numeric_limits<system_metrics::sm_pid_t>::max()) {
+    if (pid_ > static_cast<std::uint64_t>(std::numeric_limits<system_metrics::sm_pid_t>::max())) {
       fmt::print(stderr,
                  "PID value {} is out of range (maximum allowed is {}).\n",
                  pid_,
